requirements.txt
Django 
djangorestframework
django-filter
django-environ
django-autoslug 
django-countries 
Pillow 
django-phonenumber-field
phonenumbers
psycopg2-binary
flake8 
black
isort
pytest-django 
pytest-factoryboy 
Faker
pytest-cov
django-cors-headers